% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneK1K_data.R
\docType{data}
\name{cellTypeGroups}
\alias{cellTypeGroups}
\title{OneK1K cell type groups}
\format{
A named list with two entries:
\describe{
\item{B}{Vector of B cell type names ("naive B cell", "memory B cell", and "transitional stage B cell").}
\item{Mono}{Vector of monocyte type names ("CD14-positive monocyte", and "CD14-low, CD16-positive monocyte").}
}
}
\usage{
data(cellTypeGroups)
}
\description{
OneK1K cell type groups
}
\references{
\url{https://explore.data.humancellatlas.org/projects/f2078d5f-2e7d-4844-8552-f7c41a231e52}
}
\author{
Alexander Coulter \email{coultera@tamu.edu}
}
