% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneK1K_data.R
\docType{data}
\name{covariates}
\alias{covariates}
\title{OneK1K covariate data}
\format{
A data.table with 961 rows and 10 columns:
\describe{
\item{donorID}{Vector of anonymized OneK1K donor ID labels, one per cell.}
\item{age}{Vector of orthogonal polylnomial (first degree) of donor ages.}
\item{age2}{Vector of orthogonal polylnomial (second degree) of donor ages.}
\item{sex}{Vector of binary donor sex class labels (1/2).}
\item{PC_1 through PC_6}{Vectors of first 6 leading genotype principal components.}
}
}
\usage{
data(covariates)
}
\description{
OneK1K covariate data
}
\references{
url{https://explore.data.humancellatlas.org/projects/f2078d5f-2e7d-4844-8552-f7c41a231e52}
}
\author{
Alexander Coulter \email{coultera@tamu.edu}
}
