% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distQTL-package.R
\docType{package}
\name{distQTL-package}
\alias{distQTL-package}
\title{distQTL: Distribution-based eQTL mapping for scRNA-seq data}
\description{
This package conducts expression quantitative trait loci (eQTL) mapping from single cell RNA sequencing (scRNA-seq) data sets, using donor level gene expression distributions as the target responses. Inference is conducted with Fréchet regression on 2-Wasserstein space, utilizing asymptotic partial F tests developed for this context.
}
\author{
\strong{Maintainer}: Alexander Coulter \email{coultera@tamu.edu}

}
\keyword{internal}
