% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneK1K_data.R
\docType{data}
\name{geneExpression}
\alias{geneExpression}
\title{OneK1K gene expression data}
\format{
A data.table with 144,813 rows and 22 columns:
\describe{
\item{donorID}{Vector of anonymized OneK1K donor ID labels, one per cell.}
\item{cellType}{Vector of cell type labels, one per cell.}
\item{other columns}{Donor library size corrected gene expression data for 20 different genes, with gene Ensembl IDs as column names.}
}
}
\usage{
data(geneExpression)
}
\description{
OneK1K gene expression data
}
\references{
url{https://explore.data.humancellatlas.org/projects/f2078d5f-2e7d-4844-8552-f7c41a231e52}
}
\author{
Alexander Coulter \email{coultera@tamu.edu}
}
