% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneK1K_data.R
\docType{data}
\name{geneInfo}
\alias{geneInfo}
\title{OneK1K gene information}
\format{
A data.table with 20 rows and 3 columns:
\describe{
\item{geneID}{Vector of gene Ensembl IDs.}
\item{chromosome}{Vector of chromosome IDs (1-22, character).}
\item{start}{Vector of genome sequence start locations (within-chromosome), relative to hg19 reference genome.}
}
}
\usage{
data(geneInfo)
}
\description{
OneK1K gene information
}
\references{
\url{https://explore.data.humancellatlas.org/projects/f2078d5f-2e7d-4844-8552-f7c41a231e52}
}
\author{
Alexander Coulter \email{coultera@tamu.edu}
}
