% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneK1K_data.R
\docType{data}
\name{genotype}
\alias{genotype}
\title{OneK1K genotype data}
\format{
A data.table with 961 rows and 2015 columns:
\describe{
\item{donorID}{Vector of anonymized OneK1K donor ID labels.}
\item{other columns}{Donor genotype data (0/1/2) for 2014 different SNPs, with SNP names as column names.}
}
}
\usage{
data(genotype)
}
\description{
OneK1K genotype data
}
\references{
\url{https://explore.data.humancellatlas.org/projects/f2078d5f-2e7d-4844-8552-f7c41a231e52}
}
\author{
Alexander Coulter \email{coultera@tamu.edu}
}
