% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneK1K_data.R
\docType{data}
\name{snpInfo}
\alias{snpInfo}
\title{OneK1K SNP information}
\format{
A data.table with 2014 rows and 3 columns:
\describe{
\item{snpID}{Vector of SNP names as provided from OneK1K data pull.}
\item{chromosome}{Vector of chromosome IDs (1-22, character).}
\item{start}{Vector of genome sequence start locations (within-chromosome), relative to hg19 reference genome.}
}
}
\usage{
data(snpInfo)
}
\description{
OneK1K SNP information
}
\references{
url{https://explore.data.humancellatlas.org/projects/f2078d5f-2e7d-4844-8552-f7c41a231e52}
}
\author{
Alexander Coulter \email{coultera@tamu.edu}
}
