% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wasserstein_F.R
\name{Wasserstein_F}
\alias{Wasserstein_F}
\title{Wasserstein partial F-test for distribution responses.}
\usage{
Wasserstein_F(
  X,
  Y,
  lower = -Inf,
  upper = Inf,
  Q0 = NULL,
  Qm = NULL,
  C_init = NULL,
  log.p = TRUE
)
}
\arguments{
\item{X}{An $n \times p$ covariate matrix (with no intercept column), whose last column is the one being tested.}

\item{Y}{An $n \times m$ matrix row-wise containing quantile function response objects.}

\item{lower}{A numeric scalar (default \code{-Inf}) setting the lower support box constraint.}

\item{upper}{A numeric scalar (default \code{Inf}) setting the upper support box constraint.}

\item{Q0}{An optional (i.e. default \code{NULL}) $n \times m$ matrix containing the fitted quantile functions from the null Fréchet regression model, i.e. the model removing the last column of \code{X}.}

\item{Qm}{An optional (i.e. default \code{NULL}) $n \times m$ matrix containing the marginal Fréchet mean of the response quantile functions; is evaluated as \code{colMeans(Y)}.}

\item{C_init}{An optional (i.e. default \code{NULL}) $n \times (m + 1)$ matrix containing the initial Lagrangian for the QP-problem associated with the Fréchet regression problem. Positive entries correspond to active constraints; zero values correspond to inactive constraints. \code{NULL} is equivalent to specifying \code{C_init = matrix(0, n, m + 1)}.}

\item{log.p}{A boolean value (default \code{TRUE}) specifying whether the $\log(p)$-value, rather than $p$-value, should be returned from the partial F test.}
}
\value{
A list with F statistic, p-value ($\log(p)$ if specified in input), and F test degrees of freedom $df_1$ and $df_2$.
}
\description{
Wasserstein partial F-test for distribution responses.
}
